ssize again as integer (solution didn't compile in intel fortran)
git-svn-id: https://mohid.svn.codeplex.com/svn@78977 81fbe566-5dc4-48c1-bdea-7421811ca204

+++ Software/SmallTools/Convert2netcdf/ConvertToNETCDF.F90
@@ -1135,7 +1135,7 @@ program Convert2netcdf
         integer                                     :: rank, obj_type
         integer(HSIZE_T), dimension(7)              :: dims, maxdims
         integer(HID_T)                              :: gr_id, dset_id, class_id
-        integer(8)                                  :: ssize
+        integer                                     :: ssize
         integer(HID_T)                              :: space_id, datatype_id
         logical                                     :: IsMapping
         integer                                     :: ILB, IUB, JLB, JUB, KLB, KUB, i, j, k
@@ -1745,7 +1745,7 @@ if1:   if(present(Int2D) .or. present(Int3D))then
         integer                                     :: STAT_CALL
         integer(HID_T)                              :: class_id, space_id, dset_id
         integer(HID_T)                              :: datatype_id, rank, NumType
-        integer(8)                                  :: ssize
+        integer                                     :: ssize
         integer(HSIZE_T), dimension(7)              :: dims
         integer                                     :: ILB, IUB, JLB, JUB, KLB, KUB
         character(len=StringLength)                 :: Name, NCDFName, LongName, StandardName, Units
